Enhanced unit testRcvr68_92
by adding in a sparrow results file with correct results for
all 9 calibratoin options. Also added the data for this test
to work on - but had to add it both as AVLB17A_182_04:2:Rcvr68_92
and AVLB17A_182_04, since the CalSeq code is looking for the project
specified by the ScanLog.fits file.
TBD: fix this kluge.