Updated W-band unit test exp values to match GFM
Evidence points to the fact that the code for calculating
the gains from the CalSeq scan for W-band has changed in
both GFM and gbtcal - as if there was a shared external
dependency that was updated.  This results file was
derived from GFM to get the unit tests passing again.